Fix mkody/twitch-emoticons#52
If the channel using a custom name, that `name` is at the top level of the emote object, and not in `data`.
So let's shift things up to get that one (which still works even if not renamed), and then get everything else we need in `data`.

Updated tests too to try and use this case.